电机项目调整 2021-1-20 系统分主题 2021-1-20

+++ electric-machine/src/main/java/com/king/em/service/IEmProductService.java
@@ -15,7 +15,7 @@ public interface IEmProductService {
 
     boolean add(EmProduct product);
 
-    boolean delete(Long id);
+    boolean delete(Integer id);
 
     boolean update(EmProduct product);
